#25375 Move Trait logic to abstract parent class
Make use of inheritance instead of Trait logic for simpler property logic
