--- src/Traits/Objectable.php
@@ -1,22 +1,0 @@
-<?php
-/**
- * @author Jacco.Amersfoort <jacco.amersfoort@monta.nl>
- * @created 11/18/2025 12:09
- */
-namespace Monta\CheckoutApiWrapper\Traits;
-
-trait Objectable
-{
-    /**
-     * @return array
-     */
-    public function toArray(): array
-    {
-        $option = null;
-        foreach ($this as $key => $value) {
-            $option[$key] = $value;
-        }
-
-        return $option;
-    }
-}
